+++ src/game/config.ts
@@ -1,10 +1,13 @@
 import { CANVAS, Scale, Types } from 'phaser';
 import { AssetPreviewScene } from '../scenes/AssetPreviewScene';
 import { BootScene } from '../scenes/BootScene';
+import { CafeScene } from '../scenes/CafeScene';
 import { MapScene } from '../scenes/MapScene';
 import { OfficeScene } from '../scenes/OfficeScene';
+import { PoliceKioskScene } from '../scenes/PoliceKioskScene';
 import { PreloadScene } from '../scenes/PreloadScene';
 import { StreetScene } from '../scenes/StreetScene';
+import { TitleScene } from '../scenes/TitleScene';
 import { GAME_HEIGHT, GAME_WIDTH } from './constants';
 
 export const createGameConfig = (parent: string): Types.Core.GameConfig => ({
@@ -20,8 +23,11 @@ export const createGameConfig = (parent: string): Types.Core.GameConfig => ({
     scene: [
         BootScene,
         PreloadScene,
+        TitleScene,
         OfficeScene,
         StreetScene,
+        CafeScene,
+        PoliceKioskScene,
         MapScene,
         AssetPreviewScene
     ]