+++ src/scenes/CafeScene.ts
@@ -1,0 +1,7 @@
+import { StoryLocationScene } from './StoryLocationScene';
+
+export class CafeScene extends StoryLocationScene {
+    constructor() {
+        super('CafeScene', 'cafe');
+    }
+}