+++ src/scenes/PoliceKioskScene.ts
@@ -1,0 +1,7 @@
+import { StoryLocationScene } from './StoryLocationScene';
+
+export class PoliceKioskScene extends StoryLocationScene {
+    constructor() {
+        super('PoliceKioskScene', 'police-kiosk');
+    }
+}